BX:GO (Single Player) - Level 1 Mapping Complete
- finished level 1 of BX:GO
- Changed some UI.
- Fixed FirstPerson not being able to quit match in Multiplayer

+++ Source Code/Assets/Scripts/LocalScripts/LocalController.cs
@@ -145,7 +145,7 @@ public class LocalController : MonoBehaviour
         //mouse sensetivity settings
         if (PlayerPrefs.HasKey("Mouse"))
         {
-            mouseSensitivity = PlayerPrefs.GetFloat("Mouse");
+            mouseSensitivity = PlayerPrefs.GetInt("Mouse");
         }
         else
         {
